Rename to queso_unit_test.
