+++ queso/tests/b_spline_volume/test_b_spline_volume.py
@@ -1,7 +1,7 @@
 # Project imports
 from QuESo_PythonApplication.PyQuESo import PyQuESo
 from queso.python_scripts.helper import *
-from queso.python_scripts.QuESoUnittest import QuESoTestCase
+from queso.python_scripts.queso_unit_test import QuESoTestCase
 # External imports
 import unittest
 import numpy as np
