+++ queso/tests/boundary_conditions/test_boundary_conditions.py
@@ -2,7 +2,7 @@
 from QuESo_PythonApplication.PyQuESo import PyQuESo
 from QuESo_PythonApplication import MeshUtilities as MeshUtil
 from queso.python_scripts.helper import *
-from queso.python_scripts.QuESoUnittest import QuESoTestCase
+from queso.python_scripts.queso_unit_test import QuESoTestCase
 # Unittest import
 import unittest
